+++ algo/A1Z26.py
@@ -1,4 +1,4 @@
-from base.ciphers import ciphers
+from algo.ciphers import ciphers
 class A1Z26(ciphers):
     def __init__(self):
         super().__init__()
