+++ algo/alphabet.py
