+++ utils.py
@@ -1,0 +1,27 @@
+import os
+
+#var
+token = ""
+filename =""
+encoder = ""
+data =""
+global newFileName
+
+def fileIsExist(filename):
+    return os.path.exists(filename)
+
+def readFileContent(filename):
+    with open(filename, 'r') as f:
+        return f.read()
+
+def makeCopyOfFile(oldFileName, newContent, key, status = "encrypted"):
+    global newFileName
+    newFileName = status + "-"+ oldFileName
+    with open(newFileName, 'w') as f:
+        f.write(key + '\n')
+        f.write(newContent)
+    
+
+if __name__ == '__main__':
+    print("utils.py")
+
